Separated platform independent from platform dependent code in samples
