+++ VulkanSharp.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27428.1
+VisualStudioVersion = 15.0.27130.2003
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Generator", "tools\Generator\Generator.csproj", "{09F60593-810A-491C-84FB-A9848A182E6D}"
 EndProject
@@ -12,8 +12,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{102C
 		samples\Inspector.Windows\Inspector.Windows.csproj = samples\Inspector.Windows\Inspector.Windows.csproj
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClearView", "samples\ClearView\ClearView.csproj", "{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Inspector", "samples\Inspector\Inspector.csproj", "{45326935-09C1-49B0-A983-DC9249C01316}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Platforms", "Platforms", "{28D51DFD-7CC8-4408-8D62-0AFC048ACE1B}"
@@ -22,8 +20,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Vulkan.Android", "src\Platf
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Vulkan.Windows", "src\Platforms\Windows\Vulkan.Windows.csproj", "{9E914493-2F70-4F34-8BD5-48185663EC03}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XLogo", "samples\XLogo\XLogo.csproj", "{EF450800-9AE3-488D-802E-C85F226A2C37}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NativeMemory", "tests\NativeMemory\NativeMemory.csproj", "{FB79F840-669D-4B00-92C3-F23B6C24C153}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{BCBE8B1A-BDF9-4303-B113-DB8AEF369E26}"
@@ -31,6 +27,14 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DocMerge", "tools\DocMerge\DocMerge.csproj", "{3D13E652-5F08-41E0-ABB6-8802249DCB96}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Vulkan.iOS", "src\Platforms\iOS\Vulkan.iOS.csproj", "{1D44D14A-8A7A-4DD9-837D-E7F50EF2ADA6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples.Common", "samples\Samples.Common\Samples.Common.csproj", "{5C30400B-37A3-48E7-A10A-15A1CF984A9E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClearView.Android", "samples\ClearView.Android\ClearView.Android.csproj", "{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XLogo.Android", "samples\XLogo.Android\XLogo.Android.csproj", "{EF450800-9AE3-488D-802E-C85F226A2C37}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples.Android", "samples\Samples.Android\Samples.Android.csproj", "{79C2E3E0-456B-491F-962F-E424E1241763}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -46,10 +50,6 @@ Global
 		{15253E36-DDDE-4C0A-9A6E-7632BACC8F7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{15253E36-DDDE-4C0A-9A6E-7632BACC8F7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{15253E36-DDDE-4C0A-9A6E-7632BACC8F7C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Release|Any CPU.Build.0 = Release|Any CPU
 		{45326935-09C1-49B0-A983-DC9249C01316}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{45326935-09C1-49B0-A983-DC9249C01316}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45326935-09C1-49B0-A983-DC9249C01316}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -62,10 +62,6 @@ Global
 		{9E914493-2F70-4F34-8BD5-48185663EC03}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9E914493-2F70-4F34-8BD5-48185663EC03}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9E914493-2F70-4F34-8BD5-48185663EC03}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EF450800-9AE3-488D-802E-C85F226A2C37}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EF450800-9AE3-488D-802E-C85F226A2C37}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EF450800-9AE3-488D-802E-C85F226A2C37}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EF450800-9AE3-488D-802E-C85F226A2C37}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FB79F840-669D-4B00-92C3-F23B6C24C153}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FB79F840-669D-4B00-92C3-F23B6C24C153}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FB79F840-669D-4B00-92C3-F23B6C24C153}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -78,18 +74,40 @@ Global
 		{1D44D14A-8A7A-4DD9-837D-E7F50EF2ADA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1D44D14A-8A7A-4DD9-837D-E7F50EF2ADA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1D44D14A-8A7A-4DD9-837D-E7F50EF2ADA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5C30400B-37A3-48E7-A10A-15A1CF984A9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C30400B-37A3-48E7-A10A-15A1CF984A9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C30400B-37A3-48E7-A10A-15A1CF984A9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C30400B-37A3-48E7-A10A-15A1CF984A9E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{EF450800-9AE3-488D-802E-C85F226A2C37}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF450800-9AE3-488D-802E-C85F226A2C37}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF450800-9AE3-488D-802E-C85F226A2C37}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{EF450800-9AE3-488D-802E-C85F226A2C37}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF450800-9AE3-488D-802E-C85F226A2C37}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EF450800-9AE3-488D-802E-C85F226A2C37}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{79C2E3E0-456B-491F-962F-E424E1241763}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{79C2E3E0-456B-491F-962F-E424E1241763}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{79C2E3E0-456B-491F-962F-E424E1241763}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{79C2E3E0-456B-491F-962F-E424E1241763}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA} = {102C7085-1193-4401-84A7-C0A7EBD93082}
 		{45326935-09C1-49B0-A983-DC9249C01316} = {102C7085-1193-4401-84A7-C0A7EBD93082}
 		{53C04966-6CD8-456A-865B-17B03295DD06} = {28D51DFD-7CC8-4408-8D62-0AFC048ACE1B}
 		{9E914493-2F70-4F34-8BD5-48185663EC03} = {28D51DFD-7CC8-4408-8D62-0AFC048ACE1B}
-		{EF450800-9AE3-488D-802E-C85F226A2C37} = {102C7085-1193-4401-84A7-C0A7EBD93082}
 		{FB79F840-669D-4B00-92C3-F23B6C24C153} = {BCBE8B1A-BDF9-4303-B113-DB8AEF369E26}
 		{1D44D14A-8A7A-4DD9-837D-E7F50EF2ADA6} = {28D51DFD-7CC8-4408-8D62-0AFC048ACE1B}
+		{5C30400B-37A3-48E7-A10A-15A1CF984A9E} = {102C7085-1193-4401-84A7-C0A7EBD93082}
+		{D70FCB4E-05DC-46C6-B08C-86ECC1FD98CA} = {102C7085-1193-4401-84A7-C0A7EBD93082}
+		{EF450800-9AE3-488D-802E-C85F226A2C37} = {102C7085-1193-4401-84A7-C0A7EBD93082}
+		{79C2E3E0-456B-491F-962F-E424E1241763} = {102C7085-1193-4401-84A7-C0A7EBD93082}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7045423B-7050-4CA9-BD63-22FC708510BF}
