+++ samples/ClearView.Android/Assets/AboutAssets.txt
