+++ samples/ClearView.Android/MainActivity.cs
@@ -1,10 +1,12 @@
 ﻿using Android.App;
 using Android.Widget;
 using Android.OS;
+using Samples.Android;
+using Samples.Common;
 
-namespace XLogo
+namespace ClearView
 {
-	[Activity (Label = "XLogo", MainLauncher = true, Icon = "@mipmap/icon", ScreenOrientation = Android.Content.PM.ScreenOrientation.SensorPortrait)]
+	[Activity (Label = "ClearView", MainLauncher = true, Icon = "@mipmap/icon")]
 	public class MainActivity : Activity
 	{
 		protected override void OnCreate (Bundle savedInstanceState)
@@ -15,7 +17,7 @@ namespace XLogo
 			SetContentView (Resource.Layout.Main);
 
 			LinearLayout layout = FindViewById<LinearLayout> (Resource.Id.Layout);
-			layout.AddView (new XLogoView (ApplicationContext));
+			layout.AddView (new VulkanViewSample(ApplicationContext, new ClearSample()));
 		}
 	}
 }
