+++ samples/ClearView.Android/Properties/AndroidManifest.xml
