+++ samples/ClearView.Android/Properties/AssemblyInfo.cs
