+++ samples/ClearView.Android/Resources/AboutResources.txt
