+++ samples/ClearView.Android/Resources/layout/Main.axml
