+++ samples/ClearView.Android/Resources/mipmap-hdpi/Icon.png
