+++ samples/ClearView.Android/Resources/mipmap-mdpi/Icon.png
