+++ samples/ClearView.Android/Resources/mipmap-xhdpi/Icon.png
