+++ samples/ClearView.Android/Resources/mipmap-xxhdpi/Icon.png
