+++ samples/ClearView.Android/Resources/mipmap-xxxhdpi/Icon.png
