+++ samples/ClearView.Android/Resources/values/Strings.xml
