+++ samples/Samples.Common/IVulkanSample.cs
@@ -1,0 +1,8 @@
+﻿namespace Vulkan
+{
+    public interface IVulkanSample
+    {
+        void DrawFrame();
+        void Initialize(PhysicalDevice physicalDevice, SurfaceKhr surface);
+    }
+}
