+++ samples/Samples.Common/Samples.Common.csproj
@@ -1,0 +1,82 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
+    <ProjectGuid>{5C30400B-37A3-48E7-A10A-15A1CF984A9E}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <RootNamespace>Samples.Common</RootNamespace>
+    <AssemblyName>Samples.Common</AssemblyName>
+    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <TargetFrameworkProfile>Profile7</TargetFrameworkProfile>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>..\..\bin\Debug</OutputPath>
+    <DefineConstants>DEBUG;</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+    <ConsolePause>false</ConsolePause>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>full</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>..\..\bin\Release</OutputPath>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+    <ConsolePause>false</ConsolePause>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="ClearSample.cs" />
+    <Compile Include="IVulkanSample.cs" />
+    <Compile Include="Logo.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="XLogoSample.cs" />
+  </ItemGroup>
+  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
+  <PropertyGroup>
+    <BuildDependsOn>
+      BuildGeneratedSource;
+      $(BuildDependsOn)
+    </BuildDependsOn>
+  </PropertyGroup>
+  <PropertyGroup>
+    <Runtime Condition="'$(OS)' != 'Windows_NT'">mono --debug</Runtime>
+  </PropertyGroup>
+  <UsingTask TaskName="DownloadURL" TaskFactory="CodeTaskFactory" AssemblyFile="$(MSBuildToolsPath)\Microsoft.Build.Tasks.v4.0.dll">
+    <ParameterGroup>
+      <URL ParameterType="System.String" Required="true" />
+      <OutputFile ParameterType="System.String" Required="true" />
+    </ParameterGroup>
+    <Task>
+      <Reference Include="System" />
+      <Code Type="Fragment" Language="cs"><![CDATA[
+            new System.Net.WebClient ().DownloadFile (URL, OutputFile);
+        ]]></Code>
+    </Task>
+  </UsingTask>
+  <Target Name="DownloadSpecification" Condition="'$(OS)' == 'Windows_NT'">
+    <DownloadURL URL="https://raw.githubusercontent.com/KhronosGroup/Vulkan-Docs/1.0/src/spec/vk.xml" OutputFile="$(OutputPath)\vk.xml" />
+  </Target>
+  <Target Name="BuildGeneratedSource" DependsOnTargets="DownloadSpecification" Inputs="$(OutputPath)\vk.xml;$(OutputPath)\Generator.exe" Outputs="Vulkan\Enums.cs">
+    <MakeDir Directories="$(OutputPath)" />
+    <Exec Command="$(Runtime) &quot;$(OutputPath)\Generator.exe&quot; $(OutputPath)\vk.xml ." />
+  </Target>
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Vulkan\Vulkan.csproj">
+      <Project>{15253e36-ddde-4c0a-9a6e-7632bacc8f7c}</Project>
+      <Name>Vulkan</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Shaders\Shader.frag" />
+    <EmbeddedResource Include="Shaders\shader.frag.spv" />
+    <None Include="Shaders\Shader.vert" />
+    <EmbeddedResource Include="Shaders\shader.vert.spv" />
+  </ItemGroup>
+</Project>
