+++ samples/Samples.Common/Shaders/Shader.frag
