+++ samples/Samples.Common/Shaders/Shader.vert
