+++ samples/Samples.Common/Shaders/shader.frag.spv
