+++ samples/Samples.Common/Shaders/shader.vert.spv
