+++ samples/XLogo.Android/Assets/AboutAssets.txt
