+++ samples/XLogo.Android/Properties/AndroidManifest.xml
