+++ samples/XLogo.Android/Properties/AssemblyInfo.cs
