+++ samples/XLogo.Android/Resources/AboutResources.txt
