+++ samples/XLogo.Android/Resources/layout/Main.axml
