+++ samples/XLogo.Android/Resources/mipmap-hdpi/Icon.png
