+++ samples/XLogo.Android/Resources/mipmap-mdpi/Icon.png
