+++ samples/XLogo.Android/Resources/mipmap-xhdpi/Icon.png
