+++ samples/XLogo.Android/Resources/mipmap-xxhdpi/Icon.png
