+++ samples/XLogo.Android/Resources/mipmap-xxxhdpi/Icon.png
