+++ samples/XLogo.Android/Resources/values/Strings.xml
